# [WIP]Migrate the fuser code to be fully async