--- a/Cargo.toml
+++ b/Cargo.toml
@@ -32,6 +32,7 @@ env_logger = "0.8"
 clap = "2.32"
 bincode = "1.3.1"
 serde = {version = "1.0.102", features=["std", "derive"]}
+dashmap = "4.0.2"
 
 [build-dependencies]
 pkg-config = {version = "0.3.14", optional = true }