--- a/examples/null.rs
+++ b/examples/null.rs
@@ -5,8 +5,11 @@ struct NullFS;
 
 impl Filesystem for NullFS {}
 
-fn main() {
+#[tokio::main]
+async fn main() {
     env_logger::init();
     let mountpoint = env::args_os().nth(1).unwrap();
-    fuser::mount2(NullFS, mountpoint, &[MountOption::AutoUnmount]).unwrap();
+    fuser::mount2(NullFS, 5, mountpoint, &[MountOption::AutoUnmount])
+        .await
+        .unwrap();
 }