--- a/src/io_ops/blocking_io.rs
+++ b/src/io_ops/blocking_io.rs
@@ -2,7 +2,7 @@ use super::FileDescriptorRawHandle;
 use async_trait::async_trait;
 use libc::{self, c_int, c_void, size_t};
 use log::error;
-use std::{io, sync::Arc};
+use std::{io, sync::Arc, time::Duration};
 
 #[derive(Debug, Clone)]
 pub struct SubChannel {
@@ -14,7 +14,7 @@ impl SubChannel {
         &self.fd
     }
 
-    pub fn new(fd: FileDescriptorRawHandle) -> io::Result<SubChannel> {
+    pub fn new(fd: FileDescriptorRawHandle, _max_poll_timeout: Duration) -> io::Result<SubChannel> {
         Ok(SubChannel { fd: Arc::new(fd) })
     }