--- a/src/session.rs
+++ b/src/session.rs
@@ -19,12 +19,12 @@ use std::{
 use tokio::{sync::Mutex, task::JoinHandle};
 
 use crate::channel::{self, Channel};
+use crate::io_ops::ArcSubChannel;
 use crate::ll::fuse_abi as abi;
 use crate::request::Request;
 use crate::Filesystem;
 #[cfg(not(feature = "libfuse"))]
 use crate::MountOption;
-use crate::io_ops::ArcSubChannel;
 use std::ops::DerefMut;
 
 /// The max size of write requests from the kernel. The absolute minimum is 4k,