Exploring migrating the API/code to be async friendly.

Due to the nature of it, without some heavy features usage i'm not sure one can not have the tokio dependency.

This is really an exploration/thing for thoughts and feedback. Making it all async here is a pretty large change alas, so might not make sense to merge.

needs more clean up regardless.