@ianoc 2021-02-09
Been using https://github.com/filebench/filebench to explore the performance/behavior of this,

Current config is something like:
```
set mode quit alldone
set $dir=/tmp/benchmark_fuse/
set $nfiles=400
set $meandirwidth=20
set $nthreads=30
set $size1=16384k

define fileset name=bigfileset, path=$dir, size=$size1, entries=$nfiles, dirwidth=$meandirwidth, prealloc=80

define process name=fileserver,instances=2
{
        thread name=fileserverthread, memsize=50m, instances=$nthreads
        {
                flowop createfile name=createfile1,filesetname=bigfileset,fd=1
                flowop writewholefile name=wrtfile1,srcfd=1,fd=1,iosize=1m,filesetname=bigfileset
                flowop closefile name=closefile1,fd=1
                flowop deletefile name=deletefile1,fd=1,filesetname=bigfileset

                flowop openfile name=openfile1,filesetname=bigfileset,fd=3
                flowop appendfilerand name=appendfilerand1,iosize=16k,fd=3
                flowop closefile name=closefile2,fd=3

                flowop readwholefile filesetname=bigfileset,name=readfile1,iosize=1m,fd=2,opennext

                flowop finishoncount name=finish, value=30000
        }
}

```

Benchmarks are comparing this personality across fuser with xmp migrated to fuser from https://github.com/zargony/fuse-rs/pull/141

The async code in here

stackfs -- https://github.com/sbu-fsl/fuse-stackfs

And raw file system.


Comparison is done here via an SSD on my machine:


```



RUST


180.643: Run took 174 seconds...
180.646: Per-Operation Breakdown
finish               6673ops       38ops/s   0.0mb/s    0.000ms/op [0.000ms - 0.001ms]
readfile1            6733ops       39ops/s 616.3mb/s    4.185ms/op [0.952ms - 269.908ms]
closefile2           6673ops       38ops/s   0.0mb/s    1.415ms/op [0.001ms - 145.007ms]
appendfilerand1      6673ops       38ops/s   0.3mb/s    2.787ms/op [0.017ms - 248.646ms]
openfile1            6673ops       38ops/s   0.0mb/s    0.787ms/op [0.036ms - 55.331ms]
deletefile1          6605ops       38ops/s   0.0mb/s    3.150ms/op [1.641ms - 17.411ms]
closefile1           6673ops       38ops/s   0.0mb/s    0.013ms/op [0.002ms - 5.760ms]
wrtfile1             6673ops       38ops/s 613.6mb/s  455.436ms/op [52.915ms - 1726.729ms]
createfile1          6673ops       38ops/s   0.0mb/s    0.755ms/op [0.068ms - 44.040ms]
180.646: IO Summary: 53376 ops 306.740 ops/s 39/77 rd/wr 1230.1mb/s 58.576ms/op
180.646: Shutting down processes


RUST_ASYNC


64.775: Run took 46 seconds...
64.780: Per-Operation Breakdown
finish               4469ops       97ops/s   0.0mb/s    0.000ms/op [0.000ms - 0.001ms]
readfile1            4529ops       98ops/s 1558.8mb/s    7.112ms/op [0.057ms - 296.907ms]
closefile2           4469ops       97ops/s   0.0mb/s    0.018ms/op [0.005ms - 6.634ms]
appendfilerand1      4469ops       97ops/s   0.8mb/s    0.856ms/op [0.023ms - 34.194ms]
openfile1            4469ops       97ops/s   0.0mb/s    4.144ms/op [0.046ms - 120.010ms]
deletefile1          4395ops       96ops/s   0.0mb/s    4.001ms/op [0.027ms - 108.894ms]
closefile1           4469ops       97ops/s   0.0mb/s    0.098ms/op [0.007ms - 19.757ms]
wrtfile1             4469ops       97ops/s 1554.2mb/s   31.839ms/op [11.631ms - 197.638ms]
createfile1          4469ops       97ops/s   0.0mb/s    4.209ms/op [0.095ms - 141.379ms]
64.780: IO Summary: 35738 ops 776.812 ops/s 98/194 rd/wr 3113.8mb/s 6.541ms/op
64.780: Shutting down processes


STACK_FS_NO_SPLICE


62.421: Run took 44 seconds...
62.423: Per-Operation Breakdown
finish               4496ops      102ops/s   0.0mb/s    0.000ms/op [0.000ms - 0.001ms]
readfile1            4556ops      104ops/s 1640.5mb/s    7.767ms/op [0.048ms - 407.050ms]
closefile2           4496ops      102ops/s   0.0mb/s    0.663ms/op [0.010ms - 40.810ms]
appendfilerand1      4496ops      102ops/s   0.8mb/s    0.778ms/op [0.019ms - 38.119ms]
openfile1            4496ops      102ops/s   0.0mb/s    3.298ms/op [0.032ms - 149.472ms]
deletefile1          4422ops      100ops/s   0.0mb/s    6.810ms/op [0.306ms - 111.047ms]
closefile1           4496ops      102ops/s   0.0mb/s    0.805ms/op [0.012ms - 51.996ms]
wrtfile1             4496ops      102ops/s 1634.7mb/s   36.651ms/op [11.429ms - 249.419ms]
createfile1          4496ops      102ops/s   0.0mb/s    5.252ms/op [0.065ms - 148.200ms]
62.423: IO Summary: 35954 ops 817.040 ops/s 104/204 rd/wr 3276.0mb/s 7.755ms/op
62.423: Shutting down processes


STACK_FS_SPLICE


43.404: Run took 25 seconds...
43.408: Per-Operation Breakdown
finish               4278ops      171ops/s   0.0mb/s    0.000ms/op [0.000ms - 0.001ms]
readfile1            4338ops      174ops/s 2744.5mb/s    7.797ms/op [0.033ms - 341.664ms]
closefile2           4278ops      171ops/s   0.0mb/s    0.431ms/op [0.010ms - 38.044ms]
appendfilerand1      4278ops      171ops/s   1.3mb/s    0.565ms/op [0.019ms - 28.630ms]
openfile1            4278ops      171ops/s   0.0mb/s    2.515ms/op [0.038ms - 106.473ms]
deletefile1          4205ops      168ops/s   0.0mb/s    7.047ms/op [0.026ms - 113.189ms]
closefile1           4278ops      171ops/s   0.0mb/s    0.545ms/op [0.016ms - 30.744ms]
wrtfile1             4278ops      171ops/s 2737.7mb/s   29.499ms/op [9.590ms - 186.023ms]
createfile1          4278ops      171ops/s   0.0mb/s    3.711ms/op [0.070ms - 111.412ms]
43.408: IO Summary: 34211 ops 1368.319 ops/s 174/342 rd/wr 5483.5mb/s 6.515ms/op
43.408: Shutting down processes


RAW


16.356: Run took 13 seconds...
16.358: Per-Operation Breakdown
finish               6677ops      514ops/s   0.0mb/s    0.000ms/op [0.000ms - 0.014ms]
readfile1            6737ops      518ops/s 8244.0mb/s   13.104ms/op [0.004ms - 119.153ms]
closefile2           6677ops      514ops/s   0.0mb/s    0.010ms/op [0.001ms - 33.004ms]
appendfilerand1      6677ops      514ops/s   4.0mb/s    0.541ms/op [0.003ms - 62.801ms]
openfile1            6677ops      514ops/s   0.0mb/s    0.053ms/op [0.004ms - 58.258ms]
deletefile1          6616ops      509ops/s   0.0mb/s    7.698ms/op [1.436ms - 89.598ms]
closefile1           6677ops      514ops/s   0.0mb/s    0.013ms/op [0.001ms - 13.353ms]
wrtfile1             6677ops      514ops/s 8216.9mb/s   24.210ms/op [4.767ms - 171.393ms]
createfile1          6677ops      514ops/s   0.0mb/s    0.220ms/op [0.011ms - 39.077ms]
16.358: IO Summary: 53415 ops 4108.380 ops/s 518/1027 rd/wr 16464.9mb/s 5.737ms/op
16.358: Shutting down processes
```