@ianoc 2021-02-12
Great, look forward to it thanks! 

Playing around some more to try work on making the benchmarks more reproducible has been interesting:

but default features in the crate means very small bump in perf -- due to the low write size, i think its basically a huge number of serial tiny 4k writes just swamp everything, so little gains to be had. Enabling `abi-7-31` (though i think just 7-10 would do looking  at the code) yields :

(This is using cgroups to assign 2 logical cores to filebench and 3 logical cores to the client process be it rust or c code. hyper threading enabled/workstation so little fuzzyness on the exact cpu allocation/usage -- done to make numbers more stable.)
```
master: 74.340: IO Summary:  5028 ops 73.936 ops/s 15/15 rd/wr 3794.4mb/s 105.050ms/op
async: 53.401: IO Summary:  5023 ops 143.501 ops/s 29/29 rd/wr 7364.4mb/s 42.985ms/op
async second run for stability: 49.378: IO Summary:  5023 ops 162.021 ops/s 33/32 rd/wr 8314.9mb/s 46.103ms/op
stackfs: 53.201: IO Summary:  5028 ops 143.646 ops/s 29/29 rd/wr 7371.9mb/s 42.525ms/op
raw: 18.132: IO Summary:  5023 ops 334.839 ops/s 67/67 rd/wr 17183.9mb/s 19.829ms/op
```

the blocksize for the operation has a huge effect on the efficiency seemingly of the async one, i'm guessing this is just the serial latency through the kernel is the limiting factor all told, but wouldn't be 100% on that.

[results_set_short.txt](https://github.com/cberner/fuser/files/5973152/results_set_short.txt)